move the TaskName definition
